pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slide_autoencoder"
version = "0.1.0"
description = "Autoencoder architectures for histopathology image denoising"
requires-python = ">=3.8"
dependencies = [
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "Pillow>=8.0.0",
    "matplotlib>=3.3.0",
    "tqdm>=4.60.0",
    "scikit-image>=0.18.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
